refactor: desktop screens package
Switch desktop screens package as subpackage of desktop module (now it is on core package)
