+++ composeApp/src/jvmMain/kotlin/com/isuponev/tutordb/desktop/views/screens/IncomesScreen.kt
@@ -1,4 +1,4 @@
-package com.isuponev.tutordb.core.views.screens
+package com.isuponev.tutordb.desktop.views.screens
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -6,16 +6,17 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import com.isuponev.tutordb.core.logging.appLogger
 import com.isuponev.tutordb.core.resources.SharedResources
+import com.isuponev.tutordb.core.views.screens.Screen
 
-object SubjectsScreen : Screen(
-    SharedResources.strings.routeOfSubjectsScreen.localized(),
+object IncomesScreen : Screen(
+    SharedResources.strings.routeOfIncomesScreen.localized(),
 ) {
     @Composable
     override fun view(
         navHostController: NavHostController,
         modifier: Modifier
     ) {
-        appLogger.i(tag = SubjectsScreen::class.java.simpleName) { "Load subjects screen" }
-        Text(SharedResources.strings.screenSubjects.localized())
+        appLogger.i(tag = IncomesScreen::class.java.simpleName) { "Load incomes screen" }
+        Text(SharedResources.strings.screenIncomes.localized())
     }
 }
