+++ composeApp/src/jvmMain/kotlin/com/isuponev/tutordb/desktop/views/screens/LessonsScreen.kt
@@ -1,4 +1,4 @@
-package com.isuponev.tutordb.core.views.screens
+package com.isuponev.tutordb.desktop.views.screens
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -6,16 +6,17 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import com.isuponev.tutordb.core.logging.appLogger
 import com.isuponev.tutordb.core.resources.SharedResources
+import com.isuponev.tutordb.core.views.screens.Screen
 
-object SettingsScreen : Screen(
-    SharedResources.strings.routeOfSettingsScreen.localized(),
+object LessonsScreen : Screen(
+    SharedResources.strings.routeOfLessonsScreen.localized(),
 ) {
     @Composable
     override fun view(
         navHostController: NavHostController,
         modifier: Modifier
     ) {
-        appLogger.i(tag = SettingsScreen::class.java.simpleName) { "Load settings screen" }
-        Text(SharedResources.strings.screenSettings.localized())
+        appLogger.i(tag = LessonsScreen::class.java.simpleName) { "Load lessons screen" }
+        Text(SharedResources.strings.screenLessons.localized())
     }
 }
