+++ composeApp/src/jvmMain/kotlin/com/isuponev/tutordb/desktop/views/screens/SubjectsScreen.kt
@@ -1,4 +1,4 @@
-package com.isuponev.tutordb.core.views.screens
+package com.isuponev.tutordb.desktop.views.screens
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -6,16 +6,17 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import com.isuponev.tutordb.core.logging.appLogger
 import com.isuponev.tutordb.core.resources.SharedResources
+import com.isuponev.tutordb.core.views.screens.Screen
 
-object StudentsScreen : Screen(
-    SharedResources.strings.routeOfStudentsScreen.localized(),
+object SubjectsScreen : Screen(
+    SharedResources.strings.routeOfSubjectsScreen.localized(),
 ) {
     @Composable
     override fun view(
         navHostController: NavHostController,
         modifier: Modifier
     ) {
-        appLogger.i(tag = StudentsScreen::class.java.simpleName) { "Load students screen" }
-        Text(SharedResources.strings.screenStudents.localized())
+        appLogger.i(tag = SubjectsScreen::class.java.simpleName) { "Load subjects screen" }
+        Text(SharedResources.strings.screenSubjects.localized())
     }
 }
